soc-copilot/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx>=0.27.0
psycopg2-binary>=2.9.9
pydantic==2.6.4
